Cancel recents animation when touch home key(2/2)
There should send an opening task transition to remote transition
handler to finish the recents animation. For now there is no opening
transition send to Shell because the exist home activity won't be
collected while start home activity.

To collect the home activity, do not set transient launch when start
recents animation if the top activity is home, so the home activity
would be paused. Then when user touch home key to cancel recents, the
home activity will be resumed so it can be collect to the transition.

When receive opening home activity while recents is running, enter
home and dismiss recents.

Bug: 207297486
Test: 1. Enable shell transition.
      2. Setup 3rd-party launcher as default home.
      3. Entering Recents from home.
      4. "adb shell input keyevent KEYCODE_HOME", verify recents
	 animation will be dismissed.
Test: atest NexusLauncherTests:com.android.quickstep.FallbackRecentsTest
Change-Id: I689032d1fa18aa9a923aaf89077dbd73c09721b7
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>